build: enable Biome noImportCycles and break existing cycles
Turn on the `suspicious/noImportCycles` rule (error, ignoreTypes) so
runtime import cycles fail the build. Type-only cycles are erased at
compile time, so `ignoreTypes` keeps them from being flagged.

The rule is project-domain, which switches on Biome's whole-project
scanner. Exclude two file kinds it can't parse: the binary MPEG-TS test
fixtures under rs/moq-mux (their `.ts` extension collides with
TypeScript) and `.svg` assets (which Biome routes to its HTML parser).

Break the 13 pre-existing runtime cycles the rule surfaced:

- js/watch: extract BufferedRange/BufferedRanges/timeRangesToArray into a
  leaf `buffered.ts`. The cycle ran backend -> audio/video barrels ->
  their mse.ts -> back to backend for timeRangesToArray. The leaf depends
  only on @moq/net, so backend and the per-track MSE sources share it
  without importing each other.

- js/moq-boy: move the Game class, GameConfig, and KEY_MAP out of the
  entry index.ts into `game.ts`. The entry was both the public re-export
  hub and the definition home that submodules imported back into.

Public API is unchanged: the same symbols are re-exported from each
package's index.
