+++ js/watch/src/audio/decoder.ts
@@ -4,8 +4,8 @@ import * as Util from "@moq/hang/util";
 import type * as Moq from "@moq/net";
 import { Time } from "@moq/net";
 import { Effect, type Getter, Signal } from "@moq/signals";
-import type { BufferedRanges } from "../backend";
 import { base64ToBytes } from "../base64";
+import type { BufferedRanges } from "../buffered";
 import { type AudioBuffer, createAudioBuffer } from "./buffer";
 // Compiled and inlined as a blob URL via vite-plugin-worklet.
 import RenderWorklet from "./render-worklet.ts?worklet";
