+++ js/watch/src/video/mse.ts
@@ -2,8 +2,8 @@ import * as Catalog from "@moq/hang/catalog";
 import * as Container from "@moq/hang/container";
 import * as Moq from "@moq/net";
 import { Effect, type Getter, Signal } from "@moq/signals";
-import { type BufferedRanges, timeRangesToArray } from "../backend";
 import { base64ToBytes } from "../base64";
+import { type BufferedRanges, timeRangesToArray } from "../buffered";
 import type { Muxer } from "../mse";
 import type { Backend, Stats } from "./backend";
 import type { Source } from "./source";
